Release 20160303T092243
Initial commit. This release is almost the same that we had in idtracker.es (plus some extra suppor to Linux, thanks to Daniel Calovi). It is stable, and almost identical versions have been tested with all the videos in the Nature Methods paper. It is also almost identical to the compiled version that we distribute in idtracker.es.